:white_check_mark: Improve coverage
